+++ src/robowars/shared/model/Obstacle.java
@@ -6,7 +6,7 @@ public class Obstacle extends GameEntity{
 
 	private static final long serialVersionUID = 2354988184568266965L;
 	private int strength;
-	private int lifeTime;
+	private long lifeTime;
 	private boolean destroyOnContact;
 	private boolean destroyByProjectile;
 
@@ -25,10 +25,10 @@ public class Obstacle extends GameEntity{
 	public void grow(int amount){
 		if(amount>=1)this.setLength(this.getLength() + amount);
 	}
-	public void passTime(int time){
+	public void passTime(long time){
 		lifeTime += time;
 	}
-	public int getTime() {return lifeTime;}
+	public long getTime() {return lifeTime;}
 	public boolean isHittable(){return destroyOnContact;}
 	public boolean isShootable(){return destroyByProjectile;}
 	public int getStrength(){return strength;}
